+++ lib/armors.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require 'adventure/armor'
+
+require 'armors/heavy'
+require 'armors/light'
+require 'armors/medium'
+
+module Armors
+  class Error < StandardError; end
+  # Your code goes here...
+end